+++ db/hooks.php
@@ -1,0 +1,26 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+/**
+ * Hook callbacks for mod_aiknowledgecheck.
+ *
+ * @package    mod_aiknowledgecheck
+ * @copyright  2025 Essay Grader AI
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+defined('MOODLE_INTERNAL') || die();
+
+$callbacks = [
+    [
+        'hook' => \core\hook\output\before_standard_top_of_body_html_generation::class,
+        'callback' => \mod_aiknowledgecheck\hook\before_standard_top_of_body_html_generation::class
+            . '::callback',
+        'priority' => 500,
+    ],
+];